refactor: rename to reactions
