+++ reaction/reaction.go
@@ -1,4 +1,4 @@
-package action
+package reaction
 
 import (
 	"errors"
@@ -10,7 +10,7 @@ import (
 	"github.com/SecurityBrewery/catalyst/migrations"
 )
 
-func findAction(app core.App, action string) (*models.Record, bool, error) {
+func findReaction(app core.App, action string) (*models.Record, bool, error) {
 	records, err := app.Dao().FindRecordsByExpr(migrations.ReactionCollectionName, dbx.HashExp{"name": action})
 	if err != nil {
 		return nil, false, err
@@ -23,10 +23,10 @@ func findAction(app core.App, action string) (*models.Record, bool, error) {
 	return records[0], true, nil
 }
 
-func runAction(actionType, name, bootstrap, script, payload string) ([]byte, error) {
+func runReaction(actionType, name, bootstrap, script, payload string) ([]byte, error) {
 	switch actionType {
 	case "python":
-		return runPythonAction(name, bootstrap, script, payload)
+		return runPythonReaction(name, bootstrap, script, payload)
 	default:
 		return nil, errors.New("unsupported action type")
 	}
